backend/requirements.txt
alembic==1.13.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
greenlet==3.0.3
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.3
MarkupSafe==2.1.5
moviepy==1.0.3
numpy==1.26.4
pillow==10.3.0
proglog==0.1.10
python-dotenv==1.0.1
python-magic==0.4.27
requests==2.31.0
SQLAlchemy==2.0.30
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
azure-ai-textanalytics==5.3.0